test(reader): split passage chooser source guards
Why:
- Package-logic source guards should run in the fast BibleUI package test lane instead of the app-host test bundle.
- The app-host bundle should retain only coverage that genuinely needs application bootstrap.

What Changed:
- Moved reader passage chooser source/progress guard tests into BibleUITests.
- Reused package-local source lookup instead of app-host repo-root path walking.
- Updated the migration plan to reflect the remaining AppAndReader scope.

Validation:
- GitNexus impact checks for the moved test methods were LOW with no direct callers or affected flows.
- GitNexus detect_changes scope=staged reported no mapped production-flow changes.
- Ran focused BibleUI package tests for ReaderPassageChooserSourceGuardTests.
- Ran focused app-host tests for the remaining AppAndReader scene/color coverage.
- Ran git diff --cached --check, docblock guard, and source-guard checks.
- Adversarial subagent review found no issues.

Impact:
- Moves source-guard coverage to the package test lane without changing production behavior.
- Reduces app-host unit-test surface and simulator-dependent test cost.
