+++ docs/test-architecture-migration-plan.md
@@ -162,7 +162,8 @@ Blocker: `AndBibleTestSupport.swift` is a 2,233-line `extension AndBibleTests` e
 - `AndBibleTests+AppAndReader` quick-selector/document-switch behavior slice has moved to `BibleUITests`: pure quick-selector sorting/labeling/action contracts live in `BibleReaderQuickModuleSelectorTests`, and controller-level current-document switch side effects that need SWORD fixtures live in `BibleReaderDocumentSwitchControllerTests`.
 - `AndBibleTests+AppAndReader` reader source-guard slice has moved to `BibleUITests`: quick-selector toolbar source guards and the active-window native-border guard live in `ReaderSourceGuardTests`, with function-boundary extraction shared through `BibleUITestSourceLocator`.
 - `AndBibleTests+AppAndReader` reader shell construction slice has moved to `BibleUITests` because the speak mini-player, navigation drawer, and side-drawer overlay are package-owned BibleUI views with injected dependencies.
-- Current `+AppAndReader` remainder: passage chooser source/progress contracts, color helpers, and the app-hosted scene-configuration sentinel. Move each package-owned test to the lowest owning package target; keep only true app-bootstrap coverage in `AndBibleTests`.
+- `AndBibleTests+AppAndReader` passage chooser source/progress slice has moved to `BibleUITests` because those guards inspect BibleUI reader presentation source for Android full-screen chooser parity, workspace titles, native toolbar avoidance, and captured progress snapshots.
+- Current `+AppAndReader` remainder: color helpers and the app-hosted scene-configuration sentinel. Move each package-owned test to the lowest owning package target; keep only true app-bootstrap coverage in `AndBibleTests`.
 - **Split `+AppAndReader` deliberately**: `sceneConfiguration` test stays app-hosted; the `ContentView` legacy root-sidebar source scan is now a repo-standards `source-guards` check; the remaining package-owned tests should move to the lowest owning package target.
 - Run BibleUITests via `xcodebuild test -scheme BibleUITests -destination 'platform=iOS Simulator,...'` against the **committed package test scheme from Phase 0** (no app host).
 - Gate per batch: moved batch green in new target; equal count removed from app bundle; app build still green.
